[win][usb] go back to the usb code from mvme-1.0.1
This doesn't revert the whole code back but sets the defines for the usb
code to the settings they had in mvme-1.0.1: USB_WIN_USE_ASYNC 0,
USB_WIN_USE_EX_FUNCTIONS 1, USB_WIN_USE_STREAMPIPE 1

This is the code that works on our lab test machine and is in use in all
mvme versions. I still have to figure out why the mini-daq readout
doesn't seem to like it (sometimes).