remove the MVLC::trySuspendStackErrorPolling() method
The concept is not good. Callers wanting to lock the error poller using
this method can never be sure if the lock is held during the operation
they attempt next. This is the case with the DSO implementation in mvme:
the DSO GUI suspends the error poller. run_script() then tries to suspend
but doesn't get the lock. If the DSO GUI now releases its lock the stack
error poller can and will interfere with the communication done in
run_script().