improve event builder event yielding strategy
Instead of immediately outputting events even if they contain null module
data the code now checks if it is possible that a full event could be
assembled in the future, once more data has arrived. In this case the event
is held back and event building is attempted again in the next call to
buildEvents().