first attempt at dealing with non-mesytec modules in the event builder
The commit adds the concept of 'fail timestamp extraction': if the
timestamp extraction filter does produce a match on the modules data
the timestamp extraction is considered to have failed. At event building
time these timestamps now lead to an 'in_window' match result with maximum
inverse match score. This means the module events are flushed with the next
fully assembled event.