readout_parser: accept the {prefix,dynamic,suffix} readout structure again
This partly reverts changes done in
841e44778934ddfa4261bfd7b875302554556d0f (simplify readout_parser).

Being able to use the more complex readout structure is important for the
GSI use-case: inserting markers to steer their analysis.

This change makes the readout_parser accept the complex readout structure
but it does not revert back to the complex ModuleData. Instead the module
data is output as a single block containing all three parts. This means
the parser callbacks did not change. Figuring out where each part starts
and ends is now up to user code (can be done by either using markers or by
using the ModuleReadoutStructure stored in the ParserState).