make the ReadoutWorker not require SnoopQueues or StackTriggers
The parameters are now optional. If snooping is not needed no queues
have to be constructed and passed to the constructor. The same for stack
triggers: if the trigger setup is done externally no trigger values need
to be specified.

Note: regarding the stack trigger setup the readout worker can now be
asymmetric: if no stack triggers are passed to the constructor then the
stack trigger registers won't be written when starting the readout. But
when stopping all the trigger registers are currently still reset to 0!