workaround for zstd master branch issue
minizip depends on zstd and for some reason needs to check out the source
code when building on build server. The checkout fails due to the master
branch being renamed. Disabling zstd support in minizip should fix this
for now.

TODO: update minizip to a version containing the zstd branch fix