implement retries on the super/stackTransaction layers
Not sure if this is a good idea but we wanted to test it out: if a super
or stack transaction call fails it is now retried up to a total of 3
attempts. Right now all errors are treated equally and lead to starting the
next transaction attempt. This should probably be made more granular so that
we do not retry things if we are definitely disconnected from the MVLC.