readout_parser: return first encountered error from parse_readout_buffer_eth()
This means the function can now actually return errors, which it previously
only did for the ones communicated via exception.

As a side effect the "debug on error" feature in mvme now catches these.

Locating the place where e.g. a NoStackFrameFound error occurs is another
matter though and right now there is no support for it. Would have to
record offsets into the packet buffers and translate them to offsets into the
input buffer, then return or store this info somewhere.